修复无限数据元件公开容量契约

+++ src/main/java/com/fish_dan_/data_energistics/ae2/InfiniteDataCellInventory.java
@@ -13,7 +13,7 @@ import appeng.api.storage.cells.StorageCell;
 
 public final class InfiniteDataCellInventory implements StorageCell {
 
-    public static final long STORED_AMOUNT = Integer.MAX_VALUE;
+    public static final long STORED_AMOUNT = Long.MAX_VALUE;
 
     private final ItemStack stack;
